server/pyproject.toml
[project]
name = "gsf-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
